# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Coza Store API Documentation",
        "description": "Coza Store API Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/product/{product}/add-comment": {
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Add comment",
                "description": "Add a comment for a product",
                "operationId": "8b5f261ccb6c2bed640ef1024fcdde21",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "Product slug",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "blue-shirt"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "rating",
                                    "body"
                                ],
                                "properties": {
                                    "rating": {
                                        "description": "User rating for this product between 1 and 5",
                                        "type": "integer",
                                        "example": 5,
                                        "maximum": 5,
                                        "minimum": 1
                                    },
                                    "body": {
                                        "description": "User opinion about this product",
                                        "type": "string",
                                        "example": "nice"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Comment successfully registered",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Your comment has been recorded and will be displayed after review."
                                        },
                                        "data": {
                                            "properties": {
                                                "comment": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 38
                                                        },
                                                        "body": {
                                                            "type": "string",
                                                            "example": "nice"
                                                        },
                                                        "rating": {
                                                            "type": "integer",
                                                            "example": 5
                                                        },
                                                        "approved": {
                                                            "type": "boolean",
                                                            "example": false
                                                        },
                                                        "parent_id": {
                                                            "type": "integer",
                                                            "example": null,
                                                            "nullable": true
                                                        },
                                                        "created_at": {
                                                            "type": "string",
                                                            "format": "date-time",
                                                            "example": "2026-08-17T17:40:49.000000Z"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/401ResponseSchema"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/422ResponseSchema"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Too many requests",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/429ResponseSchema"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "401ResponseSchema": {
                "title": "Unauthenticated Response",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unauthenticated."
                    }
                },
                "type": "object"
            },
...
            "422ResponseSchema": {
                "title": "Validation Error Response",
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "error"
                    },
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "description": "Validation errors",
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "429ResponseSchema": {
                "title": "Too Many Requests Response",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Too many requests (throttled)"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}